Release 0 | Beta 0 | Development 0 | Experimental 5
Well, the Collada model support is still not quite finished
due to an animation problem, but hopefully it won't take too
long to find the problem and get everything working.

* Added ModelJoint
* Added ModelSkeleton
* Added ModelSkin
* Added ModelVertex
* Added ModelVertexGroup
* Added ColladaConverter

* Removed ANGeometry
* Removed ANModel

* Fixed bug: Error when no audio device was found

+++ Andor 2/src/org/andor/utils/Timer.java
@@ -130,9 +130,9 @@ public class Timer {
 	public boolean hasTimePassedHours(float hours) { return hasTimePassedMinutes(hours * 60); }
 	
 	/* The methods used to return the number of seconds, minutes and hours that has passed */
-	public float getSeconds() { return getTime() / 1000; }
-	public float getMinutes() { return getSeconds() / 60; }
-	public float getHours() { return getMinutes() / 60; }
+	public float getSeconds() { return (float) getTime() / 1000f; }
+	public float getMinutes() { return getSeconds() / 60f; }
+	public float getHours() { return getMinutes() / 60f; }
 	
 	/* The method used to return the current state of this timer */
 	public boolean isRunning() { return this.running; }
